+++ app/src/androidTest/java/com/sysadmindoc/callshield/ui/screens/more/ChangelogAccessibilityTest.kt
@@ -2,6 +2,7 @@ package com.sysadmindoc.callshield.ui.screens.more
 
 import androidx.compose.ui.test.junit4.v2.createComposeRule
 import com.sysadmindoc.callshield.ui.runStrictAccessibilityChecks
+import com.sysadmindoc.callshield.ui.setThemedContent
 import org.junit.Rule
 import org.junit.Test
 
@@ -11,7 +12,7 @@ class ChangelogAccessibilityTest {
 
     @Test
     fun changelogScreenPassesStrictAccessibilityChecks() {
-        composeRule.setContent { ChangelogScreen() }
+        composeRule.setThemedContent { ChangelogScreen() }
         composeRule.runStrictAccessibilityChecks()
     }
 }